Cleaned up some of the ScreenLock logic
